+++ document/document-format/src/lib.rs
@@ -7,19 +7,16 @@
 //! See the "On-disk container" section of `node-graph/rfcs/document-format.md` for the format spec.
 
 use std::sync::Arc;
-// `Path`, `Archive`, and `FolderBackend` are only used by the native-only path-based open/create
-// and filesystem export, so they're gated off wasm to avoid unused-import warnings.
+// `Path` and `FolderBackend` are only used by the native-only path-based open/create, so they're
+// gated off wasm to avoid unused-import warnings.
 #[cfg(not(target_family = "wasm"))]
 use std::path::Path;
 
 #[cfg(not(target_family = "wasm"))]
-use document_container::archive::Archive;
-#[cfg(not(target_family = "wasm"))]
 use document_container::backends::folder::FolderBackend;
 use document_container::{AnyContainer, AsyncContainer, ByteHolder, ContainerError};
-use graph_storage::{CommitError, Delta, HotOp, NodeMetadataSource, PeerId, Registry, Rev, Session, TimeStamp};
-use graphene_resource::ResourceFuture;
-use graphene_resource::{LoadResource, Resource, ResourceHash, ResourceStorage};
+use graph_storage::{CommitError, Delta, HotOp, NodeMetadataSource, PeerId, Registry, Session};
+use graphene_resource::{LoadResource, ResourceHash};
 
 pub mod codec;
 pub mod error;
@@ -27,15 +24,21 @@ pub mod export;
 pub mod io;
 pub mod layout;
 pub mod manifest;
+pub mod persist;
+pub mod resource;
 pub mod session_state;
 
 pub use codec::{Codec, CodecError};
 pub use error::Error;
 pub use export::{ExportFormat, ExportOptions};
 pub use io::ReadError;
-pub use layout::{GddV1, Layout};
+pub use layout::{GddV1Layout, Layout};
 pub use manifest::{Manifest, PayloadCodecs};
+pub use resource::ResourceProxy;
 pub use session_state::SessionState;
+
+/// The default [`Layout`], so callers write `GddV1` for the common `Gdd<GddV1Layout>` handle.
+pub type GddV1 = Gdd<GddV1Layout>;
 
 /// The manifest is always JSON: it is the bootstrap file, read before any other payload's codec is
 /// known, so its own codec cannot itself be configurable.
@@ -63,21 +66,21 @@ pub const DEFAULT_HOT_LOG_CODEC: Codec = Codec::MessagePackFrames;
 /// the *same* on-disk/OPFS working copy — including any writes still queued on the OPFS backend. The
 /// `Session` is cloned (a snapshot copy); the container is shared.
 #[derive(Clone)]
-pub struct Gdd<L: Layout = GddV1> {
-	session: Session,
-	working: Arc<AnyContainer>,
-	layout: L,
+pub struct Gdd<L: Layout = GddV1Layout> {
+	pub(crate) session: Session,
+	pub(crate) working: Arc<AnyContainer>,
+	pub(crate) layout: L,
 	/// In-memory copy of the manifest, kept authoritative since `Gdd` is its sole writer. Holds the
 	/// per-payload codecs (so the persist path never probes the filesystem) and `last_retired_at`
 	/// (so retirement writes the manifest without first reading it). Lets the persist path stay
 	/// fully read-free and synchronous.
-	manifest: Manifest,
+	pub(crate) manifest: Manifest,
 	/// Per-peer view settings (PTZ, rulers, etc.), persisted in `session.json` not the registry, so
 	/// they stay out of the CRDT/history. Opaque to the storage layer; the editor owns the keys/values.
-	view_settings: std::collections::HashMap<String, serde_json::Value>,
+	pub(crate) view_settings: std::collections::HashMap<String, serde_json::Value>,
 	/// Per-network view settings (node-graph nav + previewing), keyed by stable [`NetworkId`]. Same per-peer
 	/// `session.json` treatment as [`view_settings`](Self::view_settings), but scoped per network.
-	network_view_settings: std::collections::HashMap<graph_storage::NetworkId, std::collections::HashMap<String, serde_json::Value>>,
+	pub(crate) network_view_settings: std::collections::HashMap<graph_storage::NetworkId, std::collections::HashMap<String, serde_json::Value>>,
 }
 
 /// Native folder-backed convenience constructors. On wasm the editor builds an OPFS-backed
@@ -242,24 +245,6 @@ impl<L: Layout> Gdd<L> {
 		self.session.can_redo()
 	}
 
-	/// Move the undo cursor back one commit (silent-zone reflog undo) and persist the new cursor. Returns
-	/// the undone `Rev`. The working registry is rewound in place by the reverse delta, so re-snapshot it
-	/// (alongside `head`) or a reopen would read a `registry.bin` inconsistent with the persisted cursor.
-	pub fn undo(&mut self) -> Result<Rev, Error> {
-		let rev = self.session.undo()?;
-		self.persist_registry_snapshot()?;
-		self.persist_session_state()?;
-		Ok(rev)
-	}
-
-	/// Re-apply the most-recently-undone commit and persist the new cursor and re-snapshotted registry.
-	pub fn redo(&mut self) -> Result<Rev, Error> {
-		let rev = self.session.redo()?;
-		self.persist_registry_snapshot()?;
-		self.persist_session_state()?;
-		Ok(rev)
-	}
-
 	pub fn registry(&self) -> &Registry {
 		self.session.registry()
 	}
@@ -299,71 +284,6 @@ impl<L: Layout> Gdd<L> {
 		&self.manifest
 	}
 
-	/// Edit the cached manifest and persist it. Always JSON, synchronous.
-	pub fn update_manifest(&mut self, edit: impl FnOnce(&mut Manifest)) -> Result<(), Error> {
-		edit(&mut self.manifest);
-		io::write_single(&self.working, self.layout.manifest_basename(), MANIFEST_CODEC, &self.manifest)?;
-		Ok(())
-	}
-
-	/// Stage a runtime snapshot as hot ops without retiring: diff the runtime against the working
-	/// registry, append the hot frames (so a crash recovers the work), and persist proto-node
-	/// declaration bytes. The working registry reflects the edit immediately, but nothing enters durable
-	/// retired history until [`retire_pending_gesture`](Self::retire_pending_gesture). Staging on every
-	/// edit while retiring only at gesture boundaries lets several edits coalesce into one retired gesture.
-	///
-	/// # Errors
-	/// [`Error::Commit`] if the runtime diff is rejected by the session. On an [`Error::Container`] /
-	/// [`Error::Codec`] from persisting the hot frames, the session has already advanced past what the
-	/// working copy reflects, so the caller should treat the document as needing re-persist.
-	pub fn stage_runtime_snapshot<M: NodeMetadataSource>(
-		&mut self,
-		network: &graph_craft::document::NodeNetwork,
-		metadata: &M,
-		resources: &graphene_resource::ResourceRegistry,
-		byte_store: &dyn ResourceStorage,
-	) -> Result<(), Error> {
-		let (hot_ops, declaration_bytes) = self.session.stage_from_runtime(network, metadata, resources)?;
-
-		for hot_op in &hot_ops {
-			self.append_hot_frame(hot_op)?;
-		}
-
-		// Persist proto-node declaration content to the byte store (the global cache in the editor,
-		// the working-copy container for standalone export). Content-addressed, so re-storing
-		// identical bytes on every commit is an idempotent no-op.
-		for bytes in declaration_bytes.values() {
-			byte_store.store(bytes);
-		}
-		Ok(())
-	}
-
-	/// Retire every pending hot op into durable history as a single gesture (marking the batch's last
-	/// delta as the gesture boundary), then re-snapshot the registry. One gesture is one undo unit, so
-	/// the caller invokes this at each undo-step boundary and before any undo/redo. A no-op when there
-	/// are no pending hot ops.
-	pub fn retire_pending_gesture(&mut self) -> Result<Vec<Rev>, Error> {
-		let Some(up_to) = self.session.hot_log().iter().map(|hot_op| hot_op.timestamp).max() else {
-			return Ok(Vec::new());
-		};
-		self.retire_inner(up_to, true)
-	}
-
-	/// Commit a runtime snapshot as one complete gesture: stage it, then immediately retire it into
-	/// durable history. Convenience for callers that produce a whole gesture atomically (tests, and any
-	/// one-shot commit). Equivalent to [`stage_runtime_snapshot`](Self::stage_runtime_snapshot) followed
-	/// by [`retire_pending_gesture`](Self::retire_pending_gesture).
-	pub fn commit_from_runtime<M: NodeMetadataSource>(
-		&mut self,
-		network: &graph_craft::document::NodeNetwork,
-		metadata: &M,
-		resources: &graphene_resource::ResourceRegistry,
-		byte_store: &dyn ResourceStorage,
-	) -> Result<Vec<Rev>, Error> {
-		self.stage_runtime_snapshot(network, metadata, resources, byte_store)?;
-		self.retire_pending_gesture()
-	}
-
 	/// Resolve the proto-node declarations referenced by the registry into a [`graph_storage::Declarations`]
 	/// map, loading each `ProtoNode`'s bytes from `byte_store` (the global cache in the editor, the
 	/// working-copy container for standalone). Only resources referenced by `Implementation::ProtoNode`
@@ -399,166 +319,6 @@ impl<L: Layout> Gdd<L> {
 		declarations
 	}
 
-	/// Apply a hot op from the broadcast stream, appending one frame to the hot log.
-	///
-	/// # Errors
-	/// Returns [`Error::Crdt`] if the op is rejected by the session. A failure to persist the hot
-	/// frame is logged, not returned.
-	pub fn apply_hot_op(&mut self, op: HotOp) -> Result<(), Error> {
-		self.session.apply_hot_op(op.clone())?;
-		if let Err(error) = self.append_hot_frame(&op) {
-			log::error!("Failed to append hot op frame: {error}");
-		}
-		Ok(())
-	}
-
-	/// Persist freshly-staged hot ops and immediately retire them into durable history. Appends each
-	/// hot frame (so a crash before retirement still recovers the work), then retires up to the last
-	/// staged timestamp, which drains exactly these ops and re-snapshots the registry. Returns the
-	/// retired `Rev`s. A no-op when nothing was staged.
-	fn append_and_retire(&mut self, hot_ops: &[HotOp], gesture: bool) -> Result<Vec<Rev>, Error> {
-		let Some(last) = hot_ops.last() else { return Ok(Vec::new()) };
-
-		for hot_op in hot_ops {
-			self.append_hot_frame(hot_op)?;
-		}
-
-		self.retire_inner(last.timestamp, gesture)
-	}
-
-	/// Encode the history deltas identified by `revs` and append them to the history file. Iterates
-	/// `session.history()` (topological/append order) filtered by `revs` membership, so the appended
-	/// frames preserve replay order regardless of the order `revs` lists them in.
-	fn append_history_deltas(&mut self, revs: &[Rev]) -> Result<(), Error> {
-		let wanted: std::collections::HashSet<Rev> = revs.iter().copied().collect();
-		let mut buffer = Vec::new();
-		for delta in self.session.history().filter(|delta| wanted.contains(&delta.id)) {
-			self.manifest.codecs.history.append(&mut buffer, delta)?;
-		}
-		self.working.append_non_blocking(&io::path_for(self.layout.history_basename(), self.manifest.codecs.history), &buffer)?;
-		Ok(())
-	}
-
-	/// Set a local annotation (e.g. a commit message) on an existing retired delta and re-persist it.
-	/// Unlike the per-gesture marker written inline at retire, this targets an already-written delta, so
-	/// the whole history file is rewritten in topological order. O(history) — fine for occasional user
-	/// labeling, not for per-gesture marking (which uses the inline path). No-op if `rev` is unknown.
-	pub fn annotate_delta(&mut self, rev: Rev, key: &str, value: serde_json::Value) -> Result<(), Error> {
-		if self.session.annotate_delta(rev, key, value) {
-			self.rewrite_history()?;
-		}
-		Ok(())
-	}
-
-	/// Rewrite the entire history file from the in-memory session. `history()` yields deltas in
-	/// topological (append) order, which is a valid replay order, so no separate sort is needed.
-	fn rewrite_history(&mut self) -> Result<(), Error> {
-		let mut buffer = Vec::new();
-		for delta in self.session.history() {
-			self.manifest.codecs.history.append(&mut buffer, delta)?;
-		}
-		self.working.write_non_blocking(&io::path_for(self.layout.history_basename(), self.manifest.codecs.history), &buffer)?;
-		Ok(())
-	}
-
-	fn persist_session_state(&mut self) -> Result<(), Error> {
-		let state = SessionState {
-			head_rev: self.session.head_rev(),
-			redo_stack: self.session.redo_stack().to_vec(),
-			next_node_counter: self.session.next_node_counter(),
-			view_settings: self.view_settings.clone(),
-			network_view_settings: self.network_view_settings.clone(),
-		};
-		io::write_single(&self.working, self.layout.session_basename(), self.manifest.codecs.session, &state)?;
-		Ok(())
-	}
-
-	/// Re-snapshot the materialized working registry to `registry.bin`. `Session::load` trusts the stored
-	/// registry to match the persisted `head`, so any cursor move (undo/redo) that rewinds the working
-	/// registry without retiring must re-persist it or a reopen would read a registry inconsistent with
-	/// `head`. Synchronous and hot-path-safe (`write_non_blocking`).
-	fn persist_registry_snapshot(&mut self) -> Result<(), Error> {
-		io::write_single(&self.working, self.layout.registry_basename(), self.manifest.codecs.registry, self.session.registry())?;
-		Ok(())
-	}
-
-	/// The per-peer view settings read from `session.json` (PTZ, rulers, overlays, snapping, collapse).
-	/// Opaque `ui::doc::*` blobs; the editor decodes them. Empty for a fresh document.
-	pub fn view_settings(&self) -> &std::collections::HashMap<String, serde_json::Value> {
-		&self.view_settings
-	}
-
-	/// Replace the per-peer view settings and persist them to `session.json`. Called by the editor when
-	/// the viewport or a document-level toggle changes; never enters the registry, history, or CRDT.
-	pub fn set_view_settings(&mut self, view_settings: std::collections::HashMap<String, serde_json::Value>) -> Result<(), Error> {
-		self.view_settings = view_settings;
-		self.persist_session_state()
-	}
-
-	/// The per-network view settings read from `session.json` (node-graph nav + previewing), keyed by
-	/// [`NetworkId`](graph_storage::NetworkId). Opaque `ui::nav::*` / `ui::previewing` blobs the editor decodes.
-	pub fn network_view_settings(&self) -> &std::collections::HashMap<graph_storage::NetworkId, std::collections::HashMap<String, serde_json::Value>> {
-		&self.network_view_settings
-	}
-
-	/// Replace the per-network view settings and persist them to `session.json`. Per-peer, per-network; never
-	/// enters the registry, history, or CRDT.
-	pub fn set_network_view_settings(&mut self, network_view_settings: std::collections::HashMap<graph_storage::NetworkId, std::collections::HashMap<String, serde_json::Value>>) -> Result<(), Error> {
-		self.network_view_settings = network_view_settings;
-		self.persist_session_state()
-	}
-
-	fn append_hot_frame(&mut self, op: &HotOp) -> Result<(), Error> {
-		let mut buffer = Vec::new();
-		self.manifest.codecs.hot_log.append(&mut buffer, op)?;
-		self.working.append_non_blocking(&io::path_for(self.layout.hot_log_basename(), self.manifest.codecs.hot_log), &buffer)?;
-		Ok(())
-	}
-
-	/// Working-copy checkpoint: promote hot ops with timestamp `≤ up_to` into retired deltas,
-	/// append them to the history file, rewrite the hot log with remaining (unretired) ops,
-	/// re-snapshot the registry, and bump `last_retired_at` on the manifest. Synchronous.
-	pub fn retire(&mut self, up_to: TimeStamp) -> Result<Vec<Rev>, Error> {
-		self.retire_inner(up_to, false)
-	}
-
-	/// `gesture`: mark the batch's last delta as a gesture boundary (one undo unit) before its history
-	/// frame is written, so the marker persists on reopen without a later frame rewrite.
-	fn retire_inner(&mut self, up_to: TimeStamp, gesture: bool) -> Result<Vec<Rev>, Error> {
-		let new_revs = self.session.retire(up_to)?;
-
-		// Mark before `append_history_deltas` so the on-disk frame carries the boundary.
-		if gesture && let Some(&last) = new_revs.last() {
-			self.session.mark_interaction_end(last);
-		}
-
-		if !new_revs.is_empty() {
-			self.append_history_deltas(&new_revs)?;
-		}
-
-		// Rewrite hot log with whatever survived retirement.
-		let mut hot_buffer = Vec::new();
-		for hot_op in self.session.hot_log() {
-			self.manifest.codecs.hot_log.append(&mut hot_buffer, hot_op)?;
-		}
-		self.working
-			.write_non_blocking(&io::path_for(self.layout.hot_log_basename(), self.manifest.codecs.hot_log), &hot_buffer)?;
-
-		// Re-snapshot registry.
-		io::write_single(&self.working, self.layout.registry_basename(), self.manifest.codecs.registry, self.session.registry())?;
-
-		self.persist_session_state()?;
-
-		// Bump cached manifest timestamp and persist it.
-		self.update_manifest(|m| m.last_retired_at = Some(chrono::Utc::now().to_rfc3339()))?;
-
-		Ok(new_revs)
-	}
-
-	pub async fn read_resource(&self, hash: &ResourceHash) -> Result<ByteHolder, ContainerError> {
-		self.working.read(&self.layout.resource_path(hash)).await
-	}
-
 	/// Store the legacy `.graphite` document bytes verbatim inside the working copy (dual-write soak).
 	/// Synchronous (hot-path safe via `write_non_blocking`): called at the autosave boundary alongside
 	/// the registry snapshot. The bytes are opaque to `Gdd` — it never deserializes them.
@@ -570,394 +330,5 @@ impl<L: Layout> Gdd<L> {
 	/// the recovery fallback both go through here. `None` when no legacy blob was ever written.
 	pub async fn read_legacy_document(&self) -> Option<ByteHolder> {
 		self.working.read(self.layout.legacy_basename()).await.ok()
-	}
-
-	/// Register a resource under `id` and store its bytes. Commits an `AddResource` delta (a single
-	/// `DataSource::Embedded` source resolved to the content hash) through the session so the registry
-	/// records the resource and the entry replicates, then writes the bytes into the working copy's
-	/// content-addressed store. The caller owns `id` allocation.
-	pub fn add_resource(&mut self, id: graph_storage::ResourceId, bytes: &[u8]) -> Result<(), Error> {
-		let hash = ResourceHash::from(bytes);
-
-		self.working.write_non_blocking(&self.layout.resource_path(&hash), bytes)?;
-
-		let hot_ops = self.session.stage_embedded_resource(id, hash)?;
-		self.append_and_retire(&hot_ops, false)?;
-		Ok(())
-	}
-
-	/// Like [`add_resource`](Self::add_resource) but copies the bytes from a filesystem `src` rather
-	/// than buffering them. Folder backends use `fs::copy` (CoW on supported filesystems); other
-	/// backends fall back to read-then-write. Native-only: there is no filesystem source path on wasm.
-	#[cfg(not(target_family = "wasm"))]
-	pub fn add_resource_from_path(&mut self, id: graph_storage::ResourceId, hash: ResourceHash, src: &Path) -> Result<(), Error> {
-		let dest_path = self.layout.resource_path(&hash);
-		if let AnyContainer::Folder(folder) = self.working.as_ref() {
-			let full = folder.root().join(&dest_path);
-			if let Some(parent) = full.parent() {
-				std::fs::create_dir_all(parent).map_err(ContainerError::Io)?;
-			}
-			std::fs::copy(src, &full).map_err(ContainerError::Io)?;
-		} else {
-			let bytes = std::fs::read(src).map_err(ContainerError::Io)?;
-			self.working.write_non_blocking(&dest_path, &bytes)?;
-		}
-
-		let hot_ops = self.session.stage_embedded_resource(id, hash)?;
-		self.append_and_retire(&hot_ops, false)?;
-		Ok(())
-	}
-
-	pub async fn has_resource(&self, hash: &ResourceHash) -> bool {
-		self.working.exists(&self.layout.resource_path(hash)).await
-	}
-
-	pub fn remove_resource(&self, hash: &ResourceHash) -> Result<(), ContainerError> {
-		self.working.remove_non_blocking(&self.layout.resource_path(hash))
-	}
-
-	pub fn resource_proxy(&self) -> ResourceProxy<L>
-	where
-		L: Clone,
-	{
-		ResourceProxy(self.working.clone(), self.layout.clone())
-	}
-
-	/// Enumerate every resource currently in the working copy. Paths that don't parse as a
-	/// `ResourceHash` (foreign files dropped into the resources directory) are silently skipped.
-	pub async fn resource_hashes(&self) -> Result<Vec<ResourceHash>, ContainerError> {
-		let dir = self.layout.resources_dir();
-		if !self.working.list_dirs("").await?.iter().any(|d| d == dir) {
-			return Ok(Vec::new());
-		}
-		let entries = self.working.list(dir).await?;
-		let prefix = format!("{dir}/");
-		let mut hashes = Vec::with_capacity(entries.len());
-		for entry in entries {
-			let Some(name) = entry.strip_prefix(&prefix) else { continue };
-			if let Ok(hash) = name.parse::<ResourceHash>() {
-				hashes.push(hash);
-			}
-		}
-		Ok(hashes)
-	}
-
-	/// Build a self-contained export of the working copy: keeps typed payloads in their recorded
-	/// codecs (no re-encode), omits session/hot-log (peer-local + ephemeral), copies resources
-	/// straight through, then materializes as a folder, zip, or xz archive at `dest`. Does not mutate
-	/// `self` and does not buffer the full export — resources stream end-to-end. Native-only:
-	/// export writes to a filesystem path.
-	///
-	/// `byte_store` is the source for `embed_all_resources`: in the editor the working copy holds no
-	/// resource bytes (they live in the app-global cache), so embedding resolves each registry hash
-	/// through the store. It is unused when `embed_all_resources` is false.
-	///
-	/// # Errors
-	/// [`Error::InvalidExportOptions`] if the options are incoherent, or [`Error::MissingResource`] if
-	/// an embedded resource's bytes are absent from `byte_store`.
-	#[cfg(not(target_family = "wasm"))]
-	pub async fn export(&self, dest: &Path, format: ExportFormat, options: ExportOptions, byte_store: &dyn LoadResource) -> Result<(), Error> {
-		options.validate().map_err(Error::InvalidExportOptions)?;
-
-		match format {
-			ExportFormat::Folder => {
-				let mut folder = document_container::backends::folder::FolderBackend::create(dest)?;
-				let mut sink = FolderSink { folder: &mut folder };
-				self.stream_entries(options, byte_store, &mut sink).await?;
-			}
-			ExportFormat::Zip => {
-				let file = std::fs::File::create(dest).map_err(document_container::ContainerError::Io)?;
-				let mut writer = document_container::archive::Zip::writer(file)?;
-				self.stream_entries(options, byte_store, &mut writer).await?;
-				use document_container::archive::ArchiveWriter;
-				writer.finish()?;
-			}
-			ExportFormat::Xz => {
-				let file = std::fs::File::create(dest).map_err(document_container::ContainerError::Io)?;
-				let mut writer = document_container::archive::Xz::writer(file)?;
-				self.stream_entries(options, byte_store, &mut writer).await?;
-				use document_container::archive::ArchiveWriter;
-				writer.finish()?;
-			}
-		}
-
-		Ok(())
-	}
-
-	/// Build a self-contained archive of the working copy in memory and return its bytes, instead of
-	/// writing to a filesystem path. Available on every target (no `std::fs`), so the editor can hand
-	/// the bytes to the frontend to download / save. Buffers the whole archive in memory; fine for
-	/// document-sized saves, not for huge exports (the streaming `export` covers that, native-only).
-	///
-	/// `legacy_document`, when present, is embedded verbatim at `Layout::legacy_basename()`, so the
-	/// produced `.gdd` carries the legacy `.graphite` fallback the dual-write soak relies on.
-	/// `ExportFormat::Folder` has no single-file byte form and is rejected.
-	pub async fn export_to_bytes(&self, format: ExportFormat, options: ExportOptions, byte_store: &dyn LoadResource, legacy_document: Option<&[u8]>) -> Result<Vec<u8>, Error> {
-		use document_container::archive::Archive;
-
-		options.validate().map_err(Error::InvalidExportOptions)?;
-
-		let cursor = std::io::Cursor::new(Vec::new());
-		let buffer = match format {
-			ExportFormat::Folder => return Err(Error::InvalidExportOptions("folder export has no single-file byte form")),
-			ExportFormat::Zip => {
-				let mut writer = document_container::archive::Zip::writer(cursor)?;
-				self.stream_entries(options, byte_store, &mut writer).await?;
-				if let Some(legacy) = legacy_document {
-					ExportSink::write_entry(&mut writer, self.layout.legacy_basename(), legacy)?;
-				}
-				writer.finish_into()?
-			}
-			ExportFormat::Xz => {
-				let mut writer = document_container::archive::Xz::writer(cursor)?;
-				self.stream_entries(options, byte_store, &mut writer).await?;
-				if let Some(legacy) = legacy_document {
-					ExportSink::write_entry(&mut writer, self.layout.legacy_basename(), legacy)?;
-				}
-				writer.finish_into()?
-			}
-		};
-
-		Ok(buffer.into_inner())
-	}
-
-	/// Drive a sink through manifest → registry → history → resources. Payloads keep the working
-	/// copy's recorded per-payload codecs (no re-encode), so registry stays single-value and history
-	/// stays multi-value without the caller having to keep them coherent. Each entry is written one
-	/// at a time so the sink only ever sees one payload's bytes; the manifest itself is always JSON.
-	async fn stream_entries(&self, options: ExportOptions, byte_store: &dyn LoadResource, sink: &mut dyn ExportSink) -> Result<(), Error> {
-		use document_container::AsyncContainer;
-
-		let codecs = self.manifest.codecs;
-		sink.write_entry(&io::path_for(self.layout.manifest_basename(), MANIFEST_CODEC), &MANIFEST_CODEC.write_single(&self.manifest)?)?;
-
-		// Include the per-peer session state (cursor + `view_settings` like PTZ/rulers) so a `.gdd` opened on
-		// another machine restores the saved viewport and undo position. It's working-copy-only otherwise, so
-		// without this the archive's `view_settings` would be empty on open and the viewport would reset.
-		let session_state = SessionState {
-			head_rev: self.session.head_rev(),
-			redo_stack: self.session.redo_stack().to_vec(),
-			next_node_counter: self.session.next_node_counter(),
-			view_settings: self.view_settings.clone(),
-			network_view_settings: self.network_view_settings.clone(),
-		};
-		sink.write_entry(&io::path_for(self.layout.session_basename(), codecs.session), &codecs.session.write_single(&session_state)?)?;
-
-		// Hashes the working copy already holds on disk; their bytes are copied through verbatim below
-		// and don't need the byte store.
-		let working_copy_hashes: std::collections::HashSet<ResourceHash> = self.resource_hashes().await?.into_iter().collect();
-
-		// Decide which resources travel as bytes. A resource already marked `Embedded` always has its
-		// bytes materialized (in the editor they live in the byte store, not the working copy, so a
-		// plain export must still pull them). `embed_all_resources` additionally promotes link-only
-		// resources (`Url`/`FilePath`/`Font`) by prepending an `Embedded` source for a self-contained
-		// export. Bytes already in the working copy are skipped here; the copy-through pass writes them.
-		let mut export_session = self.session.clone();
-		let mut hashes_from_store: Vec<ResourceHash> = Vec::new();
-		let mut links_to_promote: Vec<graph_storage::ResourceId> = Vec::new();
-		for (id, entry) in &export_session.registry().resources {
-			let Some(hash) = entry.hash else { continue };
-			let embed = entry.has_embedded_source() || options.embed_all_resources;
-			if !embed {
-				continue;
-			}
-			if !entry.has_embedded_source() {
-				links_to_promote.push(*id);
-			}
-			if !working_copy_hashes.contains(&hash) {
-				hashes_from_store.push(hash);
-			}
-		}
-
-		// Multiple resource entries can resolve to the same content hash; dedup so each is loaded once.
-		hashes_from_store.sort_unstable();
-		hashes_from_store.dedup();
-
-		// Load the gap from the byte store (fail fast if an embedded resource is missing), then commit
-		// the link promotions as real `AddSource` deltas on the clone so the exported registry and
-		// history stay consistent. The live `Gdd` is untouched.
-		let mut embedded_bytes: Vec<(ResourceHash, Resource)> = Vec::new();
-		for hash in hashes_from_store {
-			let Some(resource) = byte_store.load(hash).await else {
-				return Err(Error::MissingResource(hash));
-			};
-			embedded_bytes.push((hash, resource));
-		}
-		export_session.embed_resource_sources(links_to_promote)?;
-
-		if options.include_registry {
-			sink.write_entry(
-				&io::path_for(self.layout.registry_basename(), codecs.registry),
-				&codecs.registry.write_single(export_session.registry())?,
-			)?;
-		}
-
-		if options.include_history {
-			let mut buffer = Vec::new();
-			for delta in export_session.history() {
-				codecs.history.append(&mut buffer, delta)?;
-			}
-			if !buffer.is_empty() {
-				sink.write_entry(&io::path_for(self.layout.history_basename(), codecs.history), &buffer)?;
-			}
-		}
-
-		// Copy whatever resource bytes the working copy already holds, tracking which hashes are
-		// covered so the embed pass doesn't re-emit them.
-		let mut emitted = std::collections::HashSet::new();
-		let resources_dir = self.layout.resources_dir();
-		if self.working.list_dirs("").await?.iter().any(|d| d == resources_dir) {
-			let prefix = format!("{resources_dir}/");
-			for path in self.working.list(resources_dir).await? {
-				if let Some(hash) = path.strip_prefix(&prefix).and_then(|name| name.parse::<ResourceHash>().ok()) {
-					emitted.insert(hash);
-				}
-				let holder = self.working.read(&path).await?;
-
-				// On native, an `External` (mmap'd) holder exposes a source path the sink can copy
-				// directly (CoW / kernel-side); other holders, and every holder on wasm (OPFS has no
-				// filesystem path), fall back to writing the in-memory bytes.
-				#[cfg(not(target_family = "wasm"))]
-				match holder.source_path() {
-					Some(src_path) => sink.write_entry_from_path(&path, src_path)?,
-					None => sink.write_entry(&path, holder.as_slice())?,
-				}
-				#[cfg(target_family = "wasm")]
-				sink.write_entry(&path, holder.as_slice())?;
-			}
-		}
-
-		// Write the embedded bytes the working copy didn't already hold.
-		for (hash, resource) in &embedded_bytes {
-			if emitted.insert(*hash) {
-				sink.write_entry(&self.layout.resource_path(hash), resource.as_ref())?;
-			}
-		}
-
-		Ok(())
-	}
-}
-
-impl<L: Layout + Send + Sync> LoadResource for Gdd<L> {
-	fn load(&self, hash: ResourceHash) -> ResourceFuture<'_> {
-		Box::pin(async move {
-			let bytes = self.working.read(&self.layout.resource_path(&hash)).await.ok()?;
-			Some(Resource::new(bytes))
-		})
-	}
-}
-pub struct ResourceProxy<T: Layout>(Arc<AnyContainer>, T);
-
-impl<L: Layout + Send + Sync> LoadResource for ResourceProxy<L> {
-	fn load(&self, hash: ResourceHash) -> ResourceFuture<'_> {
-		Box::pin(async move {
-			let bytes = self.0.read(&self.1.resource_path(&hash)).await.ok()?;
-			Some(Resource::new(bytes))
-		})
-	}
-}
-
-impl<L: Layout + Send + Sync> ResourceStorage for Gdd<L> {
-	fn store(&self, data: &[u8]) -> ResourceHash {
-		let hash = ResourceHash::from(data);
-		if let Err(error) = self.working.write_non_blocking(&self.layout.resource_path(&hash), data) {
-			log::error!("ResourceStorage::store failed for {hash}: {error}");
-		}
-		hash
-	}
-
-	fn contains(&self, hash: &ResourceHash) -> bool {
-		self.working.exists_non_blocking(&self.layout.resource_path(hash))
-	}
-
-	fn garbage_collect(&self, used: &[ResourceHash]) {
-		// `garbage_collect` is synchronous but listing resources is async, so the native path blocks on
-		// it. That's unavailable on wasm (single-threaded; `block_on` would deadlock). The editor never
-		// uses `Gdd` as the runtime `ResourceStorage` on wasm (it GCs the app-global cache instead), so
-		// this is an unreachable configuration there rather than a missing feature.
-		#[cfg(target_family = "wasm")]
-		{
-			let _ = used;
-			log::error!("ResourceStorage::garbage_collect is not supported for Gdd on wasm");
-		}
-		#[cfg(not(target_family = "wasm"))]
-		{
-			let kept: std::collections::HashSet<&ResourceHash> = used.iter().collect();
-			let hashes = match futures::executor::block_on(self.resource_hashes()) {
-				Ok(hashes) => hashes,
-				Err(error) => {
-					log::error!("Failed to list resources during garbage_collect: {error}");
-					return;
-				}
-			};
-			for hash in hashes {
-				if kept.contains(&hash) {
-					continue;
-				}
-				if let Err(error) = self.working.remove_non_blocking(&self.layout.resource_path(&hash)) {
-					log::error!("ResourceStorage::garbage_collect failed to remove {hash}: {error}");
-				}
-			}
-		}
-	}
-}
-
-/// Abstraction over the sink an export streams entries into. Lets a single async loop drive
-/// folder writes, zip writes, and xz writes without duplicating the entry sequence. The archive
-/// sinks (zip/xz) work on every target since the codecs are pure-Rust; only the folder sink and
-/// the filesystem copy-through (`write_entry_from_path`) are native-only.
-///
-/// `Send` because `stream_entries` holds `&mut dyn ExportSink` across `.await`s, so the enclosing
-/// future (e.g. the editor's save future) must be `Send` on native. The concrete sinks are all `Send`.
-trait ExportSink: Send {
-	fn write_entry(&mut self, path: &str, bytes: &[u8]) -> Result<(), Error>;
-
-	/// Copy a file from disk into the sink. Default impl reads the source into memory and
-	/// forwards to `write_entry`; sinks like the folder writer override to use `fs::copy`
-	/// (CoW on supported filesystems, kernel-side copy otherwise). Native-only: only reachable
-	/// for an `External` (mmap'd) holder, which doesn't exist on wasm.
-	#[cfg(not(target_family = "wasm"))]
-	fn write_entry_from_path(&mut self, path: &str, src: &std::path::Path) -> Result<(), Error> {
-		let bytes = std::fs::read(src).map_err(document_container::ContainerError::Io)?;
-		self.write_entry(path, &bytes)
-	}
-}
-
-#[cfg(not(target_family = "wasm"))]
-struct FolderSink<'a> {
-	folder: &'a mut document_container::backends::folder::FolderBackend,
-}
-
-#[cfg(not(target_family = "wasm"))]
-impl ExportSink for FolderSink<'_> {
-	fn write_entry(&mut self, path: &str, bytes: &[u8]) -> Result<(), Error> {
-		document_container::Container::write(self.folder, path, bytes)?;
-		Ok(())
-	}
-
-	fn write_entry_from_path(&mut self, path: &str, src: &std::path::Path) -> Result<(), Error> {
-		document_container::validate_path(path)?;
-		let dest = self.folder.root().join(path);
-		if let Some(parent) = dest.parent() {
-			std::fs::create_dir_all(parent).map_err(document_container::ContainerError::Io)?;
-		}
-		std::fs::copy(src, &dest).map_err(document_container::ContainerError::Io)?;
-		Ok(())
-	}
-}
-
-impl<W: std::io::Write + std::io::Seek + Send> ExportSink for document_container::archive::ZipWriter<W> {
-	fn write_entry(&mut self, path: &str, bytes: &[u8]) -> Result<(), Error> {
-		use document_container::archive::ArchiveWriter;
-		ArchiveWriter::write_entry(self, path, bytes)?;
-		Ok(())
-	}
-}
-
-impl<W: std::io::Write + std::io::Seek + Send> ExportSink for document_container::archive::XzWriter<W> {
-	fn write_entry(&mut self, path: &str, bytes: &[u8]) -> Result<(), Error> {
-		use document_container::archive::ArchiveWriter;
-		ArchiveWriter::write_entry(self, path, bytes)?;
-		Ok(())
 	}
 }
